additional datamodel code and everything else
more datamodel stuff was added in a not-quite-oop way. I'm sure it will be fine. Additionally, adding all other unity generated files not hidden by the .gitignore. I find it suspect that these files are included, but I will defer to those minds better than mine that built the .gitignore.